Remove the generic `feeder` field from `JsonParser`
The parser having a reference to the feeder makes it difficult to
create structs that wrap both the parser and the feeder (as this would
lead to a self-referential struct). It also  leads to the entire parser
needing to be gerneric, both over the specific type of the feeder and
the lifetime of the reference.

This change removes the generic constraints on the parser.
An exclusive reference can be passed as an argument to
`JsonParser::next_event(..)` instead. Only this method needs to be
gerneric.

Adapted test and doc-tests to comply with the change.